requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain_huggingface
langchain-openai
sentence-transformers
faiss-cpu
pypdf
python-dotenv
mlflow
groq
numpy<2
python-multipart



torch==2.2.2+cpu
torchvision==0.17.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu